feat: refactor main entry file to launch v1 and v2
