+++ conference-manager/.env.public
@@ -25,6 +25,7 @@ PRIVATE_KEY_ADMIN_V2=
 
 JWT_SECRET_KEY=unsecure_secret_key
 JWT_REFRESH_SECRET_KEY=unsecure_refresh_secret_key
+REQUIRES_AUTH=false
 
 # containers env variables: not used for deployment
 COMPOSE_PROJECT_NAME=ms-conference
