+++ conference-manager/ms-conference-api/src/infrastructure/headers.ts
@@ -1,0 +1,10 @@
+const getCORSHeaders = () => {
+  return {
+    origin: '*',
+    methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+    allowedHeaders: ['Content-Type', 'Origin', 'X-Requested-With', 'Accept', 'Authorization'],
+    exposedHeaders: ['Authorization'],
+  }
+}
+
+export { getCORSHeaders }
